+++ semcore/add-filter/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.2] - 2025-02-05
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/flex-box` [5.40.0 ~> 5.40.1], `@semcore/utils` [4.45.0 ~> 4.45.1]).
+
+## [1.1.1] - 2025-02-05
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/icon` [4.56.0 ~> 4.57.0], `@semcore/button` [5.41.0 ~> 5.41.1]).
+
 ## [1.1.0] - 2025-02-03
 
 ### Changed